+++ cartridge-cli.lua
@@ -2833,13 +2833,14 @@ local cmd_pack = {
     name = 'pack',
     doc = 'Pack application into a distributable bundle',
     usage = remove_leading_spaces([=[
-        %s pack [options] [<type>] [<path>]
+        %s pack [options] <type> [<path>]
 
         Arguments
             type                      Distribution type to create
                                       Allowed types: %s
 
             path                      Path to application
+                                      Default to current directory
 
         Options
             --name NAME               Application name
@@ -2895,7 +2896,7 @@ function cmd_pack.callback(args)
     app_state.version_release = string.format('%s-%s', version, release)
 
     -- collect pack-specific application info
-    app_state.dest_dir = fio.abspath('.')
+    app_state.dest_dir = fio.cwd()
     app_state.from = args.from
     app_state.download_token = args.download_token
     app_state.docker_build_args = args.docker_build_args
@@ -2999,6 +3000,14 @@ function cmd_pack.parse(arg)
     args.type = parameters[1]
     args.path = parameters[2]
 
+    if args.path == nil then
+        args.path = fio.cwd()
+    end
+
+    if args.type == nil then
+        die('Package type is required')
+    end
+
     if not array_contains(available_distribution_types, args.type) then
         die("Package type should be one of: %s",
                 table.concat(available_distribution_types, ', '))
@@ -3027,10 +3036,6 @@ function cmd_pack.parse(arg)
                 args.from = default_dockerfile_path
             end
         end
-    end
-
-    if args.path == nil then
-        die("Path to application is required")
     end
 
     return args
@@ -3185,7 +3190,7 @@ local function create_app_directory_and_init_git(dest_dir, template, name)
 end
 
 function cmd_create.callback(args)
-    local path = args.path or "."
+    local path = args.path and fio.abspath(args.path) or fio.cwd()
 
     if not fio.path.exists(path) then
         die("Directory doesn't exist: '%s'", path)
@@ -3235,7 +3240,7 @@ local cmd_build = {
 
         Arguments
             path                      Path to application
-                                      Default to `.`
+                                      Default to current directory
     ]=]):format(self_name),
 }
 
@@ -3246,7 +3251,7 @@ function cmd_build.parse(args)
     }
 
     if result.path == nil then
-        result.path = '.'
+        result.path = fio.cwd()
     end
 
     return result